[style] Separate file
